Fix playback UI glitch and migrate to Long for position/duration

+++ app/src/main/java/com/zeon/meplayer/ui/main/PlayingBar.kt
@@ -41,8 +41,8 @@ import com.zeon.meplayer.ui.theme.AppGradients
 fun PlayingBar(
     currentSong: Audio?,
     isPlaying: Boolean,
-    currentPosition: Int,
-    duration: Int,
+    currentPosition: Long,
+    duration: Long,
     onPlayPause: () -> Unit,
     onNext: () -> Unit,
     onPrevious: () -> Unit,
